XD_UngDung_QuanLy_StudioGame/requirements.txt
mysql-connector-python~=9.1.0
PyQt5~=5.15.11
pygubu
pandas
matplotlib
pyinstaller
Faker~=33.1.0
plotly
bcrypt~=4.2.1
PySide6~=6.8.1
tkcalendar
fpdf2
pillow
openpyxl
pytest